test(mithril-stm): move IvcProof::verify tests to proof module; document key-consistency invariant
